Fix aes cipher portability issue

+++ include/ox_aes.h
@@ -76,8 +76,8 @@
  * @details Such container should be initialized with #cx_aes_init_key_no_throw.
  */
 struct cx_aes_key_s {
-    size_t  size;      ///< key size
-    uint8_t keys[32];  ///< key value
+    uint32_t size;      ///< key size
+    uint8_t  keys[32];  ///< key value
 };
 /** Convenience type. See #cx_aes_key_s. */
 typedef struct cx_aes_key_s cx_aes_key_t;
